Code refactor into controllers
